requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Async HTTP Client
httpx>=0.24.0

# Task Queue
redis>=4.6.0

# File Processing
python-magic>=0.4.27
python-docx>=0.8.11
openpyxl>=3.1.2
python-pptx>=0.6.21
PyPDF2>=3.0.1
bleach>=6.0.0

# OpenAI Integration
openai>=1.0.0

# Security
cryptography>=40.0.1
pyjwt>=2.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.6.0
python-dateutil>=2.8.2
pytz>=2023.3

# Monitoring
prometheus-client>=0.16.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
